Dramatic changes to vmm subsystem, nothing works atm

+++ src/modules/libraries/std/strings.c
@@ -253,7 +253,7 @@ uint64_t strlen(const char *str) {
 void *memset(void *dest, int val, uint64_t size) {
     uint8_t *d = (uint8_t *)dest;
     for (uint64_t i = 0; i < size; i++) {
-        d[i] = val;
+        d[i] = (uint8_t)val;
     }
     return dest;
 }
